Add app_dir a compulsory argument and add some more checks to end the ct_test_app_dockerfile test sooner if something gets wrong

+++ test-lib.sh
@@ -820,16 +820,23 @@ ct_show_resources()
 #                        (must work with an application directory called 'app-src')
 # Argument: app_url - git URI with a testing application
 # Argument: body_regexp - PCRE regular expression that must match the response body
+# Argument: app_dir - name of the application directory that is used in the Dockerfile
 # Argument: port - Optional port number (default: 8080)
 ct_test_app_dockerfile() {
   local dockerfile=$1
   local app_url=$2
   local expected_text=$3
-  local port=${4:-8080}
-  local app_dir=app-src # this is a directory that must match with the name in the Dockerfile
+  local port=${5:-8080}
+  local app_dir=$4 # this is a directory that must match with the name in the Dockerfile
   local app_image_name=myapp
   local ret
   local cname=app_dockerfile
+
+  if ! [ -r "${dockerfile}" -a -s "${dockerfile}" ] ; then
+    echo "ERROR: Dockerfile ${dockerfile} does not exist or is empty."
+    echo "Terminating the Dockerfile build."
+    return 1
+  fi
 
   CID_FILE_DIR=${CID_FILE_DIR:-$(mktemp -d)}
   local dockerfile_abs
@@ -845,8 +852,25 @@ ct_test_app_dockerfile() {
   cat Dockerfile
 
   git clone "${app_url}" "${app_dir}"
+  if [ $? -ne 0 ] ; then
+    echo "ERROR: Git repository ${app_url} cannot be cloned into ${app_dir}."
+    echo "Terminating the Dockerfile build."
+    return 1
+  fi
+
   docker build --no-cache=true -t "${app_image_name}" .
+  if [ $? -ne 0 ] ; then
+    echo "ERROR: The image cannot be built from ${dockerfile} and application ${app_url}."
+    echo "Terminating the Dockerfile build."
+    return 1
+  fi
+
   docker run -d --cidfile="${CID_FILE_DIR}/app_dockerfile" --rm "${app_image_name}"
+  if [ $? -ne 0 ] ; then
+    echo "ERROR: The image ${app_image_name} cannot be run for ${dockerfile} and application ${app_url}."
+    echo "Terminating the Dockerfile build."
+    return 1
+  fi
   ct_wait_for_cid "${CID_FILE_DIR}/app_dockerfile"
 
   ip="$(ct_get_cip "${cname}")"
